Course 3 homework 2 corrected

+++ src/main/java/ru/hogwarts/course3/school/service/FacultyServiceImpl.java
@@ -6,6 +6,7 @@ import ru.hogwarts.course3.school.model.Faculty;
 import ru.hogwarts.course3.school.repository.FacultyRepository;
 
 import java.util.Collection;
+import java.util.NoSuchElementException;
 import java.util.stream.Collectors;
 
 @Service
@@ -19,12 +20,18 @@ public class FacultyServiceImpl implements FacultyService {
 
     @Override
     public Faculty createFaculty(Faculty faculty) {
+        if (facultyRepository.existsById(faculty.getId())) {
+            throw new IllegalArgumentException();
+        }
         return facultyRepository.save(faculty);
     }
 
     @Override
-    public Faculty findFaculty(long id) {
-        return facultyRepository.findById(id).get();
+    public Faculty findFaculty(Long id) {
+        if (facultyRepository.existsById(id)) {
+            return facultyRepository.findById(id).get();
+        }
+        return null;
     }
 
     @Override
@@ -33,9 +40,13 @@ public class FacultyServiceImpl implements FacultyService {
     }
 
     @Override
-    public void deleteFaculty(long id) {
+    public void deleteFaculty(Long id) {
+        if (!facultyRepository.existsById(id)) {
+            throw new NoSuchElementException();
+        }
         facultyRepository.deleteById(id);
     }
+
 
     @Override
     public Collection<Faculty> getAllFaculties() {
